refactor: :recycle: Refatorando e dividindo o servico puppeteer para uma outra api
