--- dirname.js
@@ -1,2 +1,0 @@
-const pathFile = __dirname;
-module.exports = pathFile;
