+++ src/assets/dirname.js
@@ -1,0 +1,2 @@
+const pathFile = __dirname;
+module.exports = pathFile;
