+++ src/routes/index.js
@@ -6,7 +6,7 @@ const errorHandler = require('../middlewares/errorHandler');
 const notImplemented = require('../middlewares/notImplemented');
 
 const router = Router();
-router.get('/', (req, res) => res.sendStatus(200));
+router.get('/', (_req, res) => res.sendStatus(200));
 router.use('/source', linker);
 router.use('/data', loader);
 router.use('/search', geter);
