+++ src/routes/linkeRouter.js
@@ -2,7 +2,6 @@ const { Router } = require('express');
 const Controller = require('../controllers/linkerController');
 
 const router = Router();
-router.get('/check', Controller.checkSource);
 router.get('/', Controller.getSource);
 
 module.exports = router;
