+++ src/utils/UnzipFile.js
@@ -1,6 +1,6 @@
 const decompress = require('decompress');
 const path = require('path');
-const pathFileAssets = require('../../dirname');
+const pathFileAssets = require('../assets/dirname');
 const fs = require('fs/promises');
 
 module.exports = async () => {
