+++ src/utils/fetchData.js
@@ -1,6 +1,6 @@
 const fs = require('fs');
 const path = require('path');
-const pathFileAssets = require('../../dirname');
+const pathFileAssets = require('../assets/dirname');
 const axios = require('axios');
 
 module.exports = async (Url) => {
@@ -16,8 +16,5 @@ module.exports = async (Url) => {
 
   console.log('Gravando o novo arquivo zip');
   const pathFile = path.resolve(pathFileAssets, 'series.zip');
-  console.log(pathFile);
-  console.log(process.cwd());
   fs.writeFileSync(pathFile, data);
-  console.log('Sucesso!');
 };
