+++ src/utils/readWrite.js
@@ -1,6 +1,6 @@
 const fs = require('fs/promises');
 const path = require('path');
-const pathFileAssets = require('../../dirname');
+const pathFileAssets = require('../assets/dirname');
 
 const readFile = async (pathFile) =>
   fs.readFile(path.resolve(pathFileAssets, pathFile), 'utf-8');
