--- vercel.json
@@ -1,15 +1,0 @@
-{
-  "version": 2,
-  "builds": [
-    {
-      "src": "src/index.js",
-      "use": "@vercel/node"
-    }
-  ],
-  "routes": [
-    {
-      "src": "(.*)",
-      "dest": "src/index.js"
-    }
-  ]
-}
